+++ containershare/validator/library.py
@@ -36,9 +36,9 @@ import os
 import re
 import sys
 from containershare.logger import bot
+from containershare.utils import read_yaml
 from glob import glob
-from .utils import notvalid
-import json
+from .utils import ( notvalid, validate_url )
 
 
 class LibraryValidator:
@@ -50,22 +50,21 @@ class LibraryValidator:
     def __str__(self):
         return "containershare.LibraryValidator"
 
-    def validate(self, expfile):
-        bot.test('CONTAINER: %s' % os.path.basename(expfile))
-        if expfile.endswith('json'):
-            if not self._validate_json(expfile):
+    def validate(self, ymlfile):
+        bot.test('CONTAINER: %s' % os.path.basename(ymlfile))
+        if ymlfile.endswith('md'):
+            if not self._validate_markdown(ymlfile):
                 return False
-        elif expfile.endswith('md'):
-            if not self._validate_markdown(expfile):
-                return False
+        else:
+            print('Invalid extension %s, %s must be markdown (.md)' % ymlfile)
+            return False
         return True
-
 
     def _print_valid(self, result):
         options = {True:'yes', False: 'no'}
         return options[result]
 
-    def _validate_markdown(self, expfile):
+    def _validate_markdown(self, ymlfile):
         '''ensure that fields are present in markdown file'''
 
         try:
@@ -74,84 +73,57 @@ class LibraryValidator:
             bot.error('Python yaml is required for testing yml/markdown files.')
             sys.exit(1)
 
-        self.metadata = {}
-        uid = os.path.basename(expfile).strip('.md')
+        uid = os.path.basename(ymlfile).strip('.md')
      
-        if os.path.exists(expfile):
-            with open(expfile, "r") as stream:
-                docs = yaml.load_all(stream)
-                for doc in docs:
-                    if isinstance(doc,dict):
-                        for k,v in doc.items():
-                            print('%s: %s' %(k,v))
-                            self.metadata[k] = v
-            self.metadata['uid'] = uid
+        if os.path.exists(ymlfile):
+
+            self.metadata = read_yaml(ymlfile)
        
-            fields = ['github', 'docker', 'web', 'name',
-                      'tags', 'uid', 'maintainer']
+            # Required fields for library
+            fields = ['github',
+                      'docker',
+                      'web',
+                      'name',
+                      'tags',
+                      'uid',
+                      'maintainer']
 
             # Tests for all fields
             for field in fields:
                 if field not in self.metadata:
+                    print('%s is missing field %s.' %(uid, field))
                     return False
                 if self.metadata[field] in ['',None]:
                     return False
 
+            # Uid in file must match filename
+            if self.metadata['uid'] != uid:
+                print('Mismatch in file %s.md and %s.' %(ymlfile, 
+                                                         self.metadata['uid'])) 
+                return False
+
+            # Ensure that urls are from Docker Hub and Github
             if 'github' not in self.metadata['github']:
                 return notvalid('%s: not a valid github repository' % name)
+            if 'hub.docker.com' not in self.metadata['docker']:
+                return notvalid('%s: not a valid github repository' % name)
+
+            # Tags must be in a list
             if not isinstance(self.metadata['tags'],list):
                 return notvalid('%s: tags must be a list' % name)
-            if not re.search("(\w+://)(.+@)*([\w\d\.]+)(:[\d]+){0,1}/*(.*)", self.metadata['github']):
-                return notvalid('%s is not a valid URL.' %(content['github']))
 
-        return True
+            # Validate that links are correct urls
+            for key in ['web', 'github', 'docker']:
+                url = self.metadata[key]
+                if not validate_url(url):
+                    return notvalid('%s is not a valid URL.' %(url))
 
+            # Validate name
+            bot.test("        Name: %s" % self.metadata['name'])
 
-    def _validate_json(self,jsonfile):
-        valids = []
-        valid = jsonfile.endswith('.json')
-        valids.append(valid)
-        bot.test("...extension json: %s" %(self._print_valid(valid))) 
-
-        with open(jsonfile,'r') as filey:
-            content = json.load(filey)
-        valid = isinstance(content,dict)
-        valids.append(valid)
-        bot.test("...json loadable: %s" %(self._print_valid(valid)))
-
-        if not all(valids):
-            return False
-
-        bot.test("Content:")
-        name = os.path.basename(jsonfile)
-        exp_id = name.replace('.json','')
-
-        # Validate name
-        if "name" not in content:
-            return notvalid('"name" not found in %s' % name)
-
-        # Library "name" corresponds to exp_id, name of repo
-        if exp_id != content['name']:
-            return notvalid('"name" not found in %s' % name)
-
-        bot.test("        Name: %s" % content['name'])
-
-        if not re.match("^[a-z0-9_-]*$", content['name']): 
-            return notvalid('''invalid characters in %s, only 
+            if not re.match("^[a-z0-9_-]*$", self.metadata['name']): 
+                return notvalid('''invalid characters in %s, only 
                                lowercase and "-" or "_" allowed.''' %(content['name'])) 
-         
-        # Validate Github
-        if "github" not in content:
-            return notvalid('"github" not found in %s' %(name)) 
-        if not re.search("(\w+://)(.+@)*([\w\d\.]+)(:[\d]+){0,1}/*(.*)",content['github']):
-            return notvalid('%s is not a valid URL.' %(content['github'])) 
-        if not isinstance(content["github"],str):
-            return notvalid("%s must be a string" %(content['github']))
-        bot.test("        Github: %s" % content['github'])
 
-        # Maintainers
-        if "maintainer" not in content:
-            return notvalid('"maintainer" missing in %s' %(name)) 
-        bot.test("        Maintainer: %s" % content['maintainer'])
 
         return True