+++ CHANGELOG.md
@@ -15,4 +15,5 @@ Critical items to know are:
 
 
 ## [v3.x](https://github.com/vsoch/containershare-python/tree/master) (master)
+ - adding Library and Runtime validator modules (0.0.10)
  - original generation of library